+++ src/main/java/uk/gov/pay/directdebit/common/clients/GoCardlessClientFactory.java
@@ -1,21 +1,23 @@
-package uk.gov.pay.directdebit.payments.clients;
+package uk.gov.pay.directdebit.common.clients;
 
 import com.gocardless.GoCardlessClient;
 import com.google.common.collect.Maps;
+import uk.gov.pay.directdebit.app.config.DirectDebitConfig;
+import uk.gov.pay.directdebit.gatewayaccounts.model.PaymentProviderAccessToken;
+import uk.gov.pay.directdebit.webhook.gocardless.config.GoCardlessFactory;
+
+import javax.net.ssl.SSLSocketFactory;
 import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.util.Map;
 import java.util.Optional;
-import javax.net.ssl.SSLSocketFactory;
-import uk.gov.pay.directdebit.app.config.DirectDebitConfig;
-import uk.gov.pay.directdebit.gatewayaccounts.model.PaymentProviderAccessToken;
-import uk.gov.pay.directdebit.webhook.gocardless.config.GoCardlessFactory;
 
 public class GoCardlessClientFactory {
 
     private final Map<PaymentProviderAccessToken, GoCardlessClientFacade> clients;
     private final SSLSocketFactory sslSocketFactory;
     private final DirectDebitConfig configuration;
+
     public GoCardlessClientFactory(
             DirectDebitConfig configuration,
             SSLSocketFactory sslSocketFactory) {
@@ -23,7 +25,7 @@ public class GoCardlessClientFactory {
         this.sslSocketFactory = sslSocketFactory;
         this.clients = Maps.newConcurrentMap();
     }
-    
+
     public GoCardlessClientFacade getClientFor(Optional<PaymentProviderAccessToken> maybeAccessToken) {
         //backward compatibility for now, will use the token in the config if it's not there
         PaymentProviderAccessToken accessToken = maybeAccessToken.orElse(PaymentProviderAccessToken.of(configuration.getGoCardless().getAccessToken()));
