+++ src/test/java/uk/gov/pay/directdebit/common/clients/GoCardlessClientFactoryTest.java
@@ -1,8 +1,6 @@
-package uk.gov.pay.directdebit.payments.clients;
+package uk.gov.pay.directdebit.common.clients;
 
 import com.gocardless.GoCardlessClient;
-import java.util.Optional;
-import javax.net.ssl.SSLSocketFactory;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -11,6 +9,9 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.pay.directdebit.app.config.DirectDebitConfig;
 import uk.gov.pay.directdebit.gatewayaccounts.model.PaymentProviderAccessToken;
+
+import javax.net.ssl.SSLSocketFactory;
+import java.util.Optional;
 
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
@@ -33,7 +34,7 @@ public class GoCardlessClientFactoryTest {
         when(mockedDirectDebitConfig.getGoCardless().getEnvironment()).thenReturn(GoCardlessClient.Environment.SANDBOX);
         goCardlessClientFactory = new GoCardlessClientFactory(mockedDirectDebitConfig, mockedSSLSocketFactory);
     }
-    
+
     @Test
     public void shouldCreateOnlyOneClientPerGatewayAccount() {
         GoCardlessClientFacade firstClient = goCardlessClientFactory
@@ -43,7 +44,7 @@ public class GoCardlessClientFactoryTest {
         assertThat(firstClient, is(secondClient));
     }
 
-    
+
     //backward compatibility, please remove once all gateway accounts have an access token
     @Test
     public void shouldCreateAClient_ifNoAccessTokenIsDefined() {
