+++ src/test/java/uk/gov/pay/directdebit/payments/services/GoCardlessServiceTest.java
@@ -4,6 +4,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.mockito.Mock;
+import uk.gov.pay.directdebit.common.clients.GoCardlessClientFacade;
+import uk.gov.pay.directdebit.common.clients.GoCardlessClientFactory;
 import uk.gov.pay.directdebit.gatewayaccounts.model.PaymentProvider;
 import uk.gov.pay.directdebit.mandate.dao.GoCardlessMandateDao;
 import uk.gov.pay.directdebit.mandate.dao.GoCardlessPaymentDao;
@@ -24,8 +26,6 @@ import uk.gov.pay.directdebit.payers.model.BankAccountDetails;
 import uk.gov.pay.directdebit.payers.model.GoCardlessBankAccountLookup;
 import uk.gov.pay.directdebit.payers.model.GoCardlessCustomer;
 import uk.gov.pay.directdebit.payers.model.SortCode;
-import uk.gov.pay.directdebit.payments.clients.GoCardlessClientFacade;
-import uk.gov.pay.directdebit.payments.clients.GoCardlessClientFactory;
 import uk.gov.pay.directdebit.payments.exception.CreateCustomerBankAccountFailedException;
 import uk.gov.pay.directdebit.payments.exception.CreateCustomerFailedException;
 import uk.gov.pay.directdebit.payments.exception.CreateMandateFailedException;
@@ -68,7 +68,7 @@ public abstract class GoCardlessServiceTest {
     protected GoCardlessClientFactory mockedGoCardlessClientFactory;
     GoCardlessService service;
 
-    GatewayAccountFixture gatewayAccountFixture =  GatewayAccountFixture
+    GatewayAccountFixture gatewayAccountFixture = GatewayAccountFixture
             .aGatewayAccountFixture().withPaymentProvider(PaymentProvider.GOCARDLESS);
     PayerFixture payerFixture = PayerFixture.aPayerFixture();
     GoCardlessCustomer goCardlessCustomer = GoCardlessCustomerFixture.aGoCardlessCustomerFixture()
@@ -76,15 +76,15 @@ public abstract class GoCardlessServiceTest {
             .withCustomerId(CUSTOMER_ID)
             .withCustomerBankAccountId(BANK_ACCOUNT_ID).toEntity();
     MandateFixture mandateFixture = aMandateFixture()
-                .withPayerFixture(payerFixture)
-                .withMandateType(MandateType.ON_DEMAND)
-                .withGatewayAccountFixture(gatewayAccountFixture)
-                .withExternalId(MANDATE_ID);
+            .withPayerFixture(payerFixture)
+            .withMandateType(MandateType.ON_DEMAND)
+            .withGatewayAccountFixture(gatewayAccountFixture)
+            .withExternalId(MANDATE_ID);
     Transaction transaction = TransactionFixture.aTransactionFixture()
             .withMandateFixture(mandateFixture)
-                .withExternalId(TRANSACTION_ID)
-                .toEntity();
-    
+            .withExternalId(TRANSACTION_ID)
+            .toEntity();
+
     GoCardlessMandate goCardlessMandate = GoCardlessMandateFixture.aGoCardlessMandateFixture().withMandateId(mandateFixture.getId()).toEntity();
     GoCardlessPayment goCardlessPayment = aGoCardlessPaymentFixture().withTransactionId(transaction.getId()).toEntity();
     BankAccountDetails bankAccountDetails = new BankAccountDetails(ACCOUNT_NUMBER, SORT_CODE);
@@ -95,9 +95,9 @@ public abstract class GoCardlessServiceTest {
         SortCode sortCode = SortCode.of("123456");
         BankAccountDetails bankAccountDetails = new BankAccountDetails(accountNumber, sortCode);
         GoCardlessBankAccountLookup goCardlessBankAccountLookup = new GoCardlessBankAccountLookup("Great Bank of Cake", true);
-        
+
         when(mockedGoCardlessClientFacade.validate(bankAccountDetails)).thenReturn(goCardlessBankAccountLookup);
-        
+
         BankAccountValidationResponse response = service.validate(mandateFixture.toEntity(), bankAccountDetails);
         assertThat(response.isValid(), is(true));
         assertThat(response.getBankName(), is("Great Bank of Cake"));
@@ -109,9 +109,9 @@ public abstract class GoCardlessServiceTest {
         SortCode sortCode = SortCode.of("123467");
         BankAccountDetails bankAccountDetails = new BankAccountDetails(accountNumber, sortCode);
         GoCardlessBankAccountLookup goCardlessBankAccountLookup = new GoCardlessBankAccountLookup(null, false);
-        
+
         when(mockedGoCardlessClientFacade.validate(bankAccountDetails)).thenReturn(goCardlessBankAccountLookup);
-        
+
         BankAccountValidationResponse response = service.validate(mandateFixture.toEntity(), bankAccountDetails);
         assertThat(response.isValid(), is(false));
         assertThat(response.getBankName(), is(nullValue()));
@@ -122,14 +122,14 @@ public abstract class GoCardlessServiceTest {
         AccountNumber accountNumber = AccountNumber.of("12345678");
         SortCode sortCode = SortCode.of("123467");
         BankAccountDetails bankAccountDetails = new BankAccountDetails(accountNumber, sortCode);
-        
+
         when(mockedGoCardlessClientFacade.validate(bankAccountDetails)).thenThrow(new RuntimeException());
-        
+
         BankAccountValidationResponse response = service.validate(mandateFixture.toEntity(), bankAccountDetails);
         assertThat(response.isValid(), is(false));
         assertThat(response.getBankName(), is(nullValue()));
     }
-    
+
     void verifyMandateFailedException() {
         when(mockedGoCardlessClientFacade.createMandate(mandateFixture.toEntity(), goCardlessCustomer)).thenThrow(new RuntimeException("gocardless said no"));
 
@@ -137,17 +137,17 @@ public abstract class GoCardlessServiceTest {
         thrown.expectMessage(format("Failed to create mandate in gocardless, mandate id: %s", MANDATE_ID));
         thrown.reportMissingExceptionWithMessage("CreateMandateFailedException expected");
     }
-    
-    void verifyCreatePaymentFailedException(){
+
+    void verifyCreatePaymentFailedException() {
         when(mockedGoCardlessClientFacade.createMandate(mandateFixture.toEntity(), goCardlessCustomer)).thenReturn(goCardlessMandate);
         when(mockedGoCardlessClientFacade.createPayment(transaction, goCardlessMandate)).thenThrow(new RuntimeException("gocardless said no"));
-        
+
         thrown.expect(CreatePaymentFailedException.class);
         thrown.expectMessage(format("Failed to create payment in gocardless, mandate id: %s, transaction id: %s",
                 MANDATE_ID, TRANSACTION_ID));
         thrown.reportMissingExceptionWithMessage("CreatePaymentFailedException expected");
     }
-    
+
     void verifyCreateCustomerBankAccountFailedException() {
         when(mockedGoCardlessClientFacade.createCustomerBankAccount(MANDATE_ID, goCardlessCustomer, payerFixture.getName(), SORT_CODE, ACCOUNT_NUMBER))
                 .thenThrow(new RuntimeException("oops"));
@@ -157,7 +157,7 @@ public abstract class GoCardlessServiceTest {
                 MANDATE_ID, payerFixture.getExternalId()));
         thrown.reportMissingExceptionWithMessage("CreateCustomerBankAccountFailedException expected");
     }
-    
+
     void verifyCreateCustomerFailedException() {
         when(mockedGoCardlessClientFacade.createCustomer(MANDATE_ID, payerFixture.toEntity()))
                 .thenThrow(new RuntimeException("ooops"));
